+++ R/corpus_micusp.R
@@ -70,7 +70,7 @@ parse_micusp_paper <- function(paperid,
 #'
 #' discipline_by_sex <-
 #'   micusp_metadata() |>
-#'   count(discipline, sex) |>
+#'   dplyr::count(discipline, sex) |>
 #'   tidyr::pivot_wider(
 #'     names_from = "sex",
 #'     values_from = "n") |>